fix : 아이디 입력 시 계속해서 네트워크 요청이 전송되던 문제 수정
