+++ frontend/src/pages/login-page/ui/components/GuestLoginForm/index.ts
@@ -1,0 +1,1 @@
+export { GuestLoginForm } from "./ui/GuestLoginForm";
